test_ssl_cert_authentication doesn't use fixed NODE_IP
