+++ tests/integration/test_ssl_cert_authentication/certs/server-ext.cnf
@@ -1,1 +1,1 @@
-subjectAltName=IP:10.5.172.77
+subjectAltName=DNS:integration-tests.clickhouse.com
